requirements.txt
Django>=2.0,<3.0
django-crispy-forms
celery==4.2.1
redis==3.2.0
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
coloredlogs==10.0
humanfriendly==4.18
idna==2.8
pyreadline==2.1
python-pushover==0.4
pyyaml>=4.2b1
requests==2.21.0
schedule==0.6.0
soupsieve==1.8
urllib3==1.24.1
twilio
django-phonenumber-field
phonenumbers
coverage==3.6